Add CUDA CRS SpMV benchmark; add /examples local CUDA_MODE define and modify benchmark harness to accommodate; switch runtime variable in /smax_benchmarks to float (still need to change the rest benchmark timers over to float); undo macros in .cu files for explicit initialization of kernel configs (need to fix)

+++ examples/benchmarks/smax_benchmarks/smax_scs_spmv.cpp
@@ -57,7 +57,7 @@ int main(int argc, char *argv[]) {
 
     // Make lambda, and pass to the benchmarking harness
     std::string bench_name = "smax_scs_spmv";
-    double runtime = 0.0;
+    float runtime = 0.0;
     int n_iter = MIN_NUM_ITERS;
     int n_threads = 1;
 #ifdef _OPENMP
